我的攻略列表

+++ PeachTravel/ToolBox/IM/Contact/ContactList/View/OptionOfFASKTableViewCell.m
@@ -40,7 +40,7 @@
         _requestNoti.hidden = YES;
     } else if (_numberOfUnreadFrendRequest > 0) {
         _requestNoti.hidden = NO;
-        _requestNoti.text = [NSString stringWithFormat:@"%ld", _numberOfUnreadFrendRequest];
+        _requestNoti.text = [NSString stringWithFormat:@"%ld", (unsigned long)_numberOfUnreadFrendRequest];
     }
 }
